requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.3.0
progressbar2==4.4.2
PyJWT==2.8.0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
